TGV Class moved to .cpp and .h
TGV functions moved from main.cpp to TGV.cpp, TGV directives moved from Headers.h to TGV.h
